Change the type of variables and parameters
Former-commit-id: e664ecc7638d2cfdba18f3ff60e1c848642e2a8d

+++ game-source-code/gui/layout/Panel.cpp
@@ -10,8 +10,8 @@ Gui::Panel::Panel(float x, float y){
 
 Dimensions Gui::Panel::getDimensions() const {
     return Dimensions{
-        static_cast<unsigned int>(panel_.getGlobalBounds().width),
-        static_cast<unsigned int>(panel_.getGlobalBounds().height)
+        panel_.getGlobalBounds().width,
+        panel_.getGlobalBounds().height
     };
 }
